+++ runtime/include/Acorn/Core/Runtime/Runtime.hpp
@@ -10,6 +10,7 @@
 #include "Acorn/Layer/LayerManager.hpp"
 #include "Acorn/Module/ModuleManager.hpp"
 #include "Acorn/Core/Version/Version.hpp"
+#include "Acorn/Threading/JobScheduler.hpp"
 
 namespace Acorn::Core
 {
@@ -26,6 +27,7 @@ namespace Acorn::Core
         void pushLayer(Args&&... args);
 
         LoggerFactory& getLoggerFactory() noexcept;
+        Threading::JobScheduler& getJobScheduler() noexcept;
 
     private:
         RuntimeAPI createAPI();
@@ -40,6 +42,7 @@ namespace Acorn::Core
 
         LoggerFactory m_loggerFactory;
         Logger m_logger;
+        Threading::JobScheduler m_jobScheduler;
         LayerManager m_layerManager;
         Module::ModuleManager m_modManager;
     };